requirements.txt
Django==2.2.17
pytz==2020.5
sqlparse==0.4.1
gunicorn==20.0.4
whitenoise==5.2.0
dj-database-url==0.5.0
django-filter==2.4.0
djangorestframework==3.12.2
Markdown==3.3.3